Let the caller manage the PIDs
This changes the signature of swptl_func_ni_init to allow multiple
different PIDs in the same unix program.

This is done by letting the caller allocate the swptl device that
represents the PID, and passing it to ni_init. Right now the device will
be freed by swptl when the last NI on it is terminated.

(Implements BRIL-1318)