Don't rely on ptl_handle_*_t types in swptl
swptl is meant to be agnostic to the portals4.h headers that are
provided, meaning that swptl.c should be able to compile with any valid
portals4.h.

The current implementation relies on the exact definition of the
ptl_handle_*_t types, this commit switches to using opaque pointers
instead.