Don't free the EQ while MDs or PTs have a reference to it
This commit disables the EQ when calling swptl_func_eq_free, but does
not free it right-away, it only marks the EQ as invalid & decrements the
reference counter.

When an event is going to be posted in an invalid EQ it is thrown, and
optionally a callback can be called to be notified.

When freeing the last structure that references the EQ the memory is
finally freed.

(Relates to BRIL-1320)