Don't rely on ptl_handle_*_t types in ME/MD parameters
If we want to avoid the reliance on the ptl_handle_*_t types like
suggested in aa5d02e we need to also avoid using the ptl_md_t/ptl_me_t
types because they embed handles.

In the future we may also need to change the `void *start` to a more
descriptive address format that allows different address spaces.