Implement seq numbers synchronization handshake
Add a synchronization handshake to update sequence numbers when
the initiator or target PID restarts. As of now, the peer had
to be restarted to restore the connection.

On initiator restarts:
On the first packet to send, the initiator sends a SYN with its
initial send sequence number. The target replies with a SYN+ACK
containing its own send sequence number. This ends the
synchronization handshake.

On target restarts:
When a packet is received (and is not a SYN), the target replies
with a NACK_RST forcing the initiator to reset its sequence
numbers and enter synchronization state (see initiator restart).

Example of a PUT after a restart of the initiator only:

  INR(restart)                 TGT
   |                            |
   |---------PUT (+SYN)-------->|
   |<---PUT_REPLY (+SYN_ACK)----|
   |-------ACK PUT_REPLY------->|
   |                            |

(Notice that the ACK of the PUT_REPLY can be piggybacked in
another packet if the initiator has other data to send.)

Example of a PUT after a restart of both the initiator and
target:

  INR(restart)             TGT(restart)
   |                            |
   |---------PUT (+SYN)-------->|
   |<-PUT_REPLY (+SYN+SYN_ACK)--|
   |--ACK PUT_REPLY (+SYN_ACK)->|
   |                            |

Example of a PUT after a restart of the target:

  INR                      TGT(restart)
   |                            |
   |-------------PUT----------->|
   |<---------NACK_RST----------|
   |                            |
 reset                          |
   |----------PUT(+SYN)-------->|
  ...