Allow configuration of UDP communications network
This patch allows to specify an IPv4 network to be used
for UDP communications thank to the BXIPKTUDP_NET environment
variable. Up to now, only localhost communications were possible.

During UDP transport initialization, the first interface in the system
having the same network prefix than BXIPKTUDP_NET value is selected.
The interface identifier part of the IP address is then used as the
Portals Network ID (nid).

For instance:
$ ip a show eth0
2: eth0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc mq state UP group default qlen 1000
    link/ether 00:15:5d:c2:7d:a1 brd ff:ff:ff:ff:ff:ff
    inet 192.168.163.90/20 brd 192.168.175.255 scope global eth0

$ BXIPKTUDP_NET="192.168.160.0" examples/get_id
GetId when the interface is physical :
Pid : 1
Nid : 858
...

The Portals Network ID is 858 (3*256+90).